requirements.txt
psycopg2-binary==2.9.9
py4j==0.10.9
pyspark==3.1.2
python-dotenv==0.21.1
